+++ java/Examples/src/main/java/com/useanvil/examples/entity/EtchCastRef.java
@@ -1,4 +1,4 @@
-package com.useanvil.examples.entity.request;
+package com.useanvil.examples.entity;
 
 
 public class EtchCastRef implements IAttachable {
